src/test/java/com/garygriffaw/itrequestservice/controllers/RequestStatusControllerTest.java
package com.garygriffaw.itrequestservice.controllers;

import com.garygriffaw.itrequestservice.config.JwtService;
import com.garygriffaw.itrequestservice.config.SecurityConfiguration;
import com.garygriffaw.itrequestservice.model.RequestStatusDTO;
import com.garygriffaw.itrequestservice.services.RequestStatusService;
import com.garygriffaw.itrequestservice.token.TokenRepository;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.context.annotation.Import;
import org.springframework.http.MediaType;
import org.springframework.security.authentication.AuthenticationProvider;
import org.springframework.security.test.context.support.WithMockUser;
import org.springframework.security.web.authentication.logout.LogoutHandler;
import org.springframework.test.web.servlet.MockMvc;

import java.util.ArrayList;
import java.util.List;

import static org.mockito.BDDMockito.given;
import static org.mockito.Mockito.times;
import static org.mockito.Mockito.verify;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@WebMvcTest(RequestStatusController.class)
@Import(SecurityConfiguration.class)
class RequestStatusControllerTest {
    @Autowired
    MockMvc mockMvc;

    @MockBean
    RequestStatusService requestStatusService;

    @MockBean
    JwtService jwtService;

    @MockBean
    AuthenticationProvider authenticationProvider;

    @MockBean
    TokenRepository tokenRepository;

    @MockBean
    LogoutHandler logoutHandler;

    @WithMockUser(username = "abc")
    @Test
    void testListRequestStatuses() throws Exception {
        List<RequestStatusDTO> testList = new ArrayList<>();

        given(requestStatusService.listRequestStatuses())
                .willReturn(testList);

        mockMvc.perform(get(RequestStatusController.REQUEST_STATUS_PATH)
                        .accept(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk())
                .andExpect(content().contentType(MediaType.APPLICATION_JSON));

        verify(requestStatusService, times(1)).listRequestStatuses();
    }
}